Promjene

+++ BACKEND/Models/Igrice.cs
@@ -5,7 +5,7 @@ namespace BACKEND.Models
     public class Igrica : Entitet
     {
         public string Naziv { get; set; } = "";
-        public decimal Ocjena { get; set; }
+        public byte Ocjena { get; set; }
         public DateTime GodinaIzdanja { get; set; }
       
     }
